make sdaccel runtime optional
