+++ src/runtime/opencl/opencl_device_api.cc
@@ -283,38 +283,9 @@ void OpenCLWorkspace::Init(const std::vector<std::string>& device_types,
   }
 }
 
-OpenCLThreadEntry* SDAccelWorkspace::GetThreadEntry() {
-  return SDAccelThreadEntry::ThreadLocal();
-}
-
-const std::shared_ptr<SDAccelWorkspace>& SDAccelWorkspace::Global() {
-  static std::shared_ptr<SDAccelWorkspace> inst = std::make_shared<SDAccelWorkspace>();
-  return inst;
-}
-
-void SDAccelWorkspace::Init() {
-  OpenCLWorkspace::Init({"accelerator"}, "Xilinx");
-}
-
-bool SDAccelWorkspace::IsOpenCLDevice(TVMContext ctx) {
-  return ctx.device_type == static_cast<DLDeviceType>(kDLSDAccel);
-}
-
-typedef dmlc::ThreadLocalStore<SDAccelThreadEntry> SDAccelThreadStore;
-
-SDAccelThreadEntry* SDAccelThreadEntry::ThreadLocal() {
-  return SDAccelThreadStore::Get();
-}
-
 TVM_REGISTER_GLOBAL("device_api.opencl")
 .set_body([](TVMArgs args, TVMRetValue* rv) {
     DeviceAPI* ptr = OpenCLWorkspace::Global().get();
-    *rv = static_cast<void*>(ptr);
-  });
-
-TVM_REGISTER_GLOBAL("device_api.sdaccel")
-.set_body([](TVMArgs args, TVMRetValue* rv) {
-    DeviceAPI* ptr = SDAccelWorkspace::Global().get();
     *rv = static_cast<void*>(ptr);
   });
